refactor: rename news feed highlight context
